[jax2tf] Create a jax_export library with JAX-only pieces for native serialization
This is a pure refactor, no functionality should change.

PiperOrigin-RevId: 518982222
