+++ jax/experimental/jax2tf/tests/jax2tf_test.py
@@ -36,6 +36,7 @@ from jax._src import test_util as jtu
 import jax._src.xla_bridge
 from jax.config import config
 from jax.experimental import jax2tf
+from jax.experimental.jax2tf import jax_export
 from jax.experimental.jax2tf.tests import tf_test_util
 from jax.experimental.maps import xmap
 from jax.experimental.shard_map import shard_map
@@ -1493,7 +1494,7 @@ class Jax2TfTest(tf_test_util.JaxToTfTestCase):
         stack.enter_context(mesh)
       # Run the JAX native version, to check it works, and to fill caches.
       _ = func_to_convert(*args)
-      exported = jax2tf.jax2tf.serialize_native(
+      exported = jax_export.serialize_native(
           func_to_convert,
           [core.ShapedArray(a.shape, a.dtype) for a in args],
           lowering_platform='tpu',
